modules/location-service-grpc/requirements.txt
GeoAlchemy2
greenlet
grpcio
grpcio-tools
kafka-python
packaging
protobuf
psycopg2-binary
SQLAlchemy